Add models and translations for advanced cells
