+++ src/main/java/appeng/datagen/AE2ItemModelProvider.java
@@ -36,5 +36,9 @@ public class AE2ItemModelProvider extends ItemModelProvider {
         basicItem(AE2Items.INSCRIBER_LOGIC_PRESS.get());
         basicItem(AE2Items.INSCRIBER_ENGINEERING_PRESS.get());
         basicItem(AE2Items.INSCRIBER_CALCULATION_PRESS.get());
+        basicItem(AE2Items.BASIC_CELL_1K.get());
+        basicItem(AE2Items.BASIC_CELL_4K.get());
+        basicItem(AE2Items.BASIC_CELL_16K.get());
+        basicItem(AE2Items.BASIC_CELL_64K.get());
     }
 }
