+++ src/main/java/appeng/items/storage/BasicCell16kItem.java
@@ -1,0 +1,9 @@
+package appeng.items.storage;
+
+public class BasicCell16kItem extends BasicCellItem {
+    private static final int CAPACITY = 16_384;
+
+    public BasicCell16kItem(Properties properties) {
+        super(properties, CAPACITY);
+    }
+}
