+++ src/main/java/appeng/items/storage/BasicCell4kItem.java
@@ -1,0 +1,9 @@
+package appeng.items.storage;
+
+public class BasicCell4kItem extends BasicCellItem {
+    private static final int CAPACITY = 4096;
+
+    public BasicCell4kItem(Properties properties) {
+        super(properties, CAPACITY);
+    }
+}
