+++ src/main/java/appeng/items/storage/BasicCell64kItem.java
@@ -1,0 +1,9 @@
+package appeng.items.storage;
+
+public class BasicCell64kItem extends BasicCellItem {
+    private static final int CAPACITY = 65_536;
+
+    public BasicCell64kItem(Properties properties) {
+        super(properties, CAPACITY);
+    }
+}
